requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
sentence-transformers>=2.2.2
pymilvus>=2.3.0

# PDF processing
reportlab>=4.0.0
PyMuPDF>=1.23.0
pdfminer.six>=20221105
pdfrw>=0.4

# Web framework and utilities
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0


# Streamlit interface
streamlit>=1.28.0

milvus==2.3.5